move thunk to top level
