+++ src/aioway/_core/__init__.py
@@ -6,5 +6,4 @@ from .iters import *
 from .nodes import *
 from .procs import *
 from .tensors import *
-from .thunks import *
 from .ufuncs import *
