+++ src/aioway/_core/tensors.py
@@ -8,8 +8,7 @@ from collections import abc as cabc
 import torch
 
 from .iters import Iter
-from .thunks import LazyThunk
-from .ufuncs import UFunc
+from .ufuncs import LazyThunk, UFunc
 
 __all__ = [
     "TensorUFunc1",
