+++ src/aioway/_core/ufuncs.py
@@ -1,12 +1,22 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
+import abc
 import typing
 from collections import abc as cabc
 
-from .iters import Iter
-from .thunks import LazyThunk
+from aioway.thunks import Thunk
 
-__all__ = ["UFunc"]
+from .iters import Iter
+from .nodes import GraphNode, node_dcls
+
+__all__ = ["UFunc", "LazyThunk"]
+
+
+@node_dcls
+class LazyThunk[T](Thunk[T], GraphNode, abc.ABC):
+    """
+    The `Thunk` that lazily invokes its dependencies, with back tracing info.
+    """
 
 
 @typing.runtime_checkable
