+++ src/aioway/modes/tracking/hists.py
@@ -10,7 +10,7 @@ from collections import abc as cabc
 
 import torch
 
-from aioway._core import TensorInput, Thunk
+from aioway._core import TensorInput
 from aioway._utils import (
     find_nested_tensors,
     is_leaf_has_grad,
@@ -18,6 +18,7 @@ from aioway._utils import (
     topo_sort,
 )
 from aioway.spaces import Attr
+from aioway.thunks import Thunk
 
 from ..modes import ModeThunk
